Improved modularity of code and now can plan motion for negative distance
